Moved Stateful/Statelesl to serde.
