+++ src/gluonts/core/serde/__init__.py
@@ -11,6 +11,7 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
+
 from ._base import encode, decode
 from ._json import dump_json, load_json
 from ._repr import dump_code, load_code
