+++ src/gluonts/core/serde/_base.py
@@ -24,7 +24,7 @@ from pydantic import BaseModel
 
 # Relative imports
 from gluonts.core import fqname_for
-from ..ty import Stateful
+
 
 bad_type_msg = textwrap.dedent(
     """
@@ -41,6 +41,39 @@ bad_type_msg = textwrap.dedent(
     """
 ).lstrip()
 
+
+
+class StatelessMeta(type):
+    def __call__(cls, *args, **kwargs):
+        self = cls.__new__(cls, *args, **kwargs)
+        if isinstance(self, cls):
+            if hasattr(self.__init__, "__checked__"):
+                (this, *args), kwargs = self.__init__.__checked__(
+                    self, *args, **kwargs
+                )
+                self.__init__.__wrapped__(this, *args, **kwargs)
+            else:
+                self.__init__(*args, **kwargs)
+            self.__init_args__ = args, kwargs
+            self.__sealed__ = True
+        return self
+
+
+class Stateless(metaclass=StatelessMeta):
+    def __getnewargs_ex__(self):
+        return self.__init_args__
+
+    def __setattr__(self, name, value):
+        if hasattr(self, "__sealed__"):
+            classname = self.__class__.__name__
+            raise ValueError(
+                f"Assignment to `{name}` outside of `{classname}.__init__`."
+            )
+        return object.__setattr__(self, name, value)
+
+
+class Stateful:
+    pass
 
 class Kind(str, Enum):
     Type = "type"
